Use proper extenders for API Serializer
Also, most of the complex logic can be moved directly to DiffSerializer
