+++ extend.php
@@ -19,10 +19,12 @@
 namespace IanM\Diff;
 
 use Flarum\Api\Serializer\BasicPostSerializer;
+use Flarum\Api\Serializer\PostSerializer;
 use Flarum\Extend;
 use Flarum\Post\Post;
 use Illuminate\Contracts\Events\Dispatcher;
 use IanM\Diff\Api\Controllers;
+use IanM\Diff\Api\SerializeDiffsOnPosts;
 use IanM\Diff\Api\Serializers\DiffSerializer;
 use IanM\Diff\Console\ArchiveCommand;
 use IanM\Diff\Models\Diff;
@@ -50,7 +52,6 @@ return [
 
 
         $events->subscribe(Listeners\PostActions::class);
-        $events->subscribe(Listeners\AddDiffRelationship::class);
 
         //$app->register(Providers\ConsoleProvider::class);
     },
@@ -60,6 +61,9 @@ return [
 
     (new Extend\ApiSerializer(BasicPostSerializer::class))
         ->hasMany('diff', DiffSerializer::class),
+
+    (new Extend\ApiSerializer(PostSerializer::class))
+        ->mutate(SerializeDiffsOnPosts::class),
 
     (new Extend\Settings())
         ->serializeToForum('textFormattingForDiffPreviews', 'the-turk-diff.textFormatting', function ($value) {
